handle Rails deprecations in tests
Dynamic :controller and :action segments have been deprecated. We replace them with a static list of routes that we use in tests.

Another deprecation is “`to_time` will always preserve the receiver timezone rather than system local time in Rails 8.1”. We don’t really care one way or another in this test app, so we opt in for the new behaviour.