Remove actionpack-action_caching from test setup
The `caches_action` call was added in one of the first commits, when we
were testing against Rails 2.3 -- long before actionpack-action_caching
was removed from Rails and published as a separate gem. A gem which has
not been updated for three years and which is compatible with Rails 6.1
and older.